Extract useStageLayout from WheelModal
The WheelModal was juggling ResizeObserver, raf-debounced measurement
and the centerOffset formula on top of phase/items/points state. Move
those layout reads and the centering math into a hook so the modal is
left with the reel orchestration. The centerOffset formula now lives in
one place with a comment explaining the optical lift and the navH floor.
